fix: restore chat history reconstruction for tool-assisted turns
